requirements.txt
pip>=19.1
colorama
opencv-python
Pillow
selenium
openpyxl
browsermob-proxy
cx-Oracle
pymysql